fixed $in list of UUID query issue --> query provider integrated

+++ protocol/src/main/java/hms/dto/GeoQuery.java
@@ -3,23 +3,23 @@ package hms.dto;
 public class GeoQuery {
 	private double latitude;
 	private double longitude;	
-	private double distance;	
+	private int distance;	
 	
 	public GeoQuery() {
 		
 	}
 	
-	public GeoQuery(double latitude, double longitude, double distance) {
+	public GeoQuery(double latitude, double longitude, int distance) {
 		this.distance = distance;
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
 	
-	public double getDistance() {
+	public int getDistance() {
 		return distance;
 	}	
 	
-	public void setDistance(double distance) {
+	public void setDistance(int distance) {
 		this.distance = distance;
 	}
